Kombinerte modaler for å bekrefte sletting og fullføring av samarbeid.
